third_party/WebKit/LayoutTests/crypto/gc-3-expected.txt

Test that window.crypto wrapper protects all dependencies, so it can always be used to create crypto.subtle.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS [object Crypto] is defined.
PASS [object SubtleCrypto] is defined.
PASS successfullyParsed is true

TEST COMPLETE
